requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
pypdf
google-cloud-aiplatform>=1.38
requests
streamlit_modal

llama_index
gitingest
markdown
fastapi
uvicorn
llama-index-llms-ollama 
llama-index-agent-openai 
llama-index-llms-openai